moved acf-json directory from the ppj theme to the ppj mu-plugin
As the ACF functionality is more to with content and less to do with presentation
it is better if it lives outside of the theme

Not only will changes made in the backend for ACF be saved in a JSON file,
changes to the JSON file will be reflected in the backend.
This should help ensure that ACF components remain exactly the same
across different environments.