created video cover image functionality
If a cover-image url argument is supplied to the video-player shortcode
then this image will be displayed until the play button is clicked
at which point the image and play button will disappear
and the video iframe will be loaded and start playing automatically.

If a cover-image is not specified then the default cover image for the iframe video will be shown.

no longer loading the wistia script E-v1.js;
removed the php template partials for raptmedia, youtube and wistia;
videoPlayer partial no creates a vue component tag;