enabled separate color themes for each leg
The particular leg that the user is visiting determines which css file is loaded.
There is one css file for each leg including the landing page.
The leg name is now a body class.
The leg name corresponds to the name of the nav menu that is displayed.
Currently each leg theme just defines the main color and the tertiary color,
and imports all of main.sass.
Color main shadow is now derived by darkening color-main by 5%.
The background color of the highlighted text blocks is
determined by color-tertiary with a 0.1 opacity
(sass functions like lighten didn't work as lightening a color
beyond a certain percentage just makes it white)