Fix issue where IE11 displayed extra spacing around quote marks in quotations
Internet Explorer is weird with the way it sized SVGs. So if a SVG isn't specifically given a height, it'll have height:auto, which IE displays as 150px. Therefore the quotation mark SVG was displaying as 150px tall, which caused extra visual spacing around it.

The fix is to specify a height on the SVG. So in addition to the width attribute, I've added height. It would be preferable for the SVG to scale by aspect ratio, but this would need further investigation.

This resource was especially useful in trying to understand how different browsers behave when scaling SVGs: https://css-tricks.com/scale-svg/