Stop header hero from showing up to the party when it isn't invited
The div `.header__hero` was still present on pages which don't use this element. However due to a recent CSS change, this element now has a minimum height – so still has a height even when it contains no content. This unwanted presence was causing it to block clicks on the content below. This became problematic on the 'Find A Job' page where the map zoom buttons sat directly underneath this ghost element.