fixed mobile nav for landing page not showing
refactored the mobile nav menu, so that it is now composed of two sections:
the primary section at the top, and the secondary section at the bottom.

On legs, the primary section will display the leg nav menu,
whereas on the landing page, the primary section will display the site wide nav menu.

On legs, the secondary section will display the site wide nav menu.