removed redundant conversion of flag to boolean value
geolocation.isBusy and searchTerm.isGeolocation are used a boolean flags.
tgeolocation.isBusy is always either a true or false boolean primitive and so
doesn't require a boolean type safety net.

searchTerm.isGeolocation is saved as a string however it is
converted to a boolean when the value is being restored
so the type safety net is not required here either.