`this.currentState` is now the single source of truth for application state
I've refactored the code to make the computed property `this.currentState` the source of truth for search and map state. A watcher on this property is now responsible for performing calls to `history.replaceState`, which means individual event listeners (e.g. list scroll, map bounds change) no longer need to know to persist the application state to the URL. They just need to update the relevant data property, which will in turn update the computed property.

The initial application state is now configured entirely within the `data()` handler, so the variable `this.urlState` is no longer passed around throughout the application. This is because URL parameters map directly to properties in the data object, so the values are read from & written to those properties.

Additionally, I've removed the 50ms timeout hack which was used in the `mounted` method to restore the scroll position of the search results list. The list scroll position must be set after the job vacancy data has been retrieved and rendered into the DOM, so this is now handled by a `$nextTick` call in the `jobs` property watcher.