Loading spinner displayed after View Job & Apply link is clicked
As it can take anywhere from 0 to up to 40+ seconds for the Justice Jobs site to load,
it was decided there should be some way of communicating to the user that their
intent to navigate to the Justice Jobs site had registered.
At the moment this is communicated by the screen dimming and a message
and a loading spinner appearing when the 'View Job & Apply' link is clicked.

Created a new Vue component ScreenOverlay.vue;
ScreenOverlay accepts an active state attribute and a message attribute.
This has a sub component Spinner.vue.

ScreenOverlay is a sub component of FindAJob.vue;
JobSummary.vue is now a sub component of FindAJob.vue.

Renamed window.ppj.screenOverlay to be window.ppj.pageContainerOverlay
- this is the overlay that is activated when the mobile menu is open.
Removed unnecessary `content` attribute for `.pageContainer__overlay`.

main.sass now imports sass components screen-overlay.sass and spinner.sass