Don't empty the `/tmp` directory on docker image build
We should not empty the `/tmp` directory when the docker image is built. The base image `mojdigital/wordpress-base` includes a file called `/tmp/keeptmp` which is used to mitigate a bug whereby `yas3fs` self-destructs by deleting the entire `/tmp` directory. This is a known issue which has been reported to the developers of `yas3fs`: https://github.com/danilop/yas3fs/issues/150

The bug here is that we were emptying the `/tmp` directory at build time, thus disabling the `yas3fs` bug workaround.