Improve reliability of 'View job' link navigation
By listening for known map/state events to complete, and then directing the browser to the job URL. This replaces the use of an arbitrary 800ms timeout which was hacky but in most cases seemed to work. By taking the time to understand the flow of events, we've been able to build a better solution.