<img> elements now replaced by background images on ie11
As this is a responsive site, the images need to resize properly.
Unfortunately ie11 does not support the object-fit property which would facilitate this.
Therefore on ie11 <img> tags are replaced with a background img on the parent element
as ie11 does support `background-size: cover` css rule.

However the parent element cannot have the `background-image` set upon
loading as this would mean every browser, not just ie, would download
a background image. Therefore the value for the background image
is stored in a data attribute `data-ie-bg-img`.
Javascript can then be used to detect if the current browser is ie11;
if so then the the value within `data-ie-bg-img` is set as
the background-image value.

A utility mixin was created to implement browser specific css;
so far only Internet Explorer 11 is supported.