Removing code to filter out unicode characters.
Mysterious LSEP characters kept appearing in the content of the site.
This was probably due to copying and pasting content from MS Word documents.

As we're now using the Visual Editor, this will automatically strip out unusual characters.
Although it won't remove the LSEP characters from the start of the line,
this happens rarely enough that it was deemed
that adding a solutino to completely solve this issue would be overkill
and might introduce new problems.