Add support for alt text on Wistia video cover images
The video shortcode has been changed to accept an attribute `cover-image-id`, which is the Media Library attachment ID of the image to use for the cover. This replaces the old attribute `cover-image` which was simply the image URL to use.

By using the WordPress function `wp_get_attachment_image` with this attachment ID, we get:
- alt text for the image
- responsive images with `srcset` and `sizes` attributes