Made images more accessible
The `alt` property has been added to <img> elements.

Where appropriate, the `aria-labelledby` property has been added to SVGs.
This property accepts the ids of a `title` and `desc` elements
that contain the title content and description content
for the element that has the `aria-labelledby` property.

If it is anticipated that the SVG will be used more than once on the page,
then the `aria-labelledby` is not used, as this property depends on using an
id property and ids must be unique for accessibility
(and it is also good HTML semantics)
An `aria-label` is used instead.

The contents of the SVGs have been tidied up:
- `id` properties have been removed
- the properties have been alphabetised
- <g> tags have been removed

The `aria-label` property has been added to the mobile nav close button.