Adding a Job Alert Message (JAM) to find-a-job page
For a period of two weeks (but perhaps longer) there will be a significantly reduced number of jobs on the site.
During this time, we would like to capture the email addresses of users that end up on the find-a-job page
so that we can notify them in the future when more jobs are available.

The JAM will only display for a given leg's find-a-job page, if the JAM has been activated in the backend
using the relevant check box AND the 'Job alert activate date' is in the past.

If the above conditions are met and the 'Job alert deactivate date' is not set,
then the JAM will display indefinitely.
However if the 'Job alert deactivate date' is in the past, then the JAM will not be displayed,
regardless of the above settings.

The text of the JAM can be specified by the relevant WYSIWIG field on the 'find-a-job' page settings.