Adding leg page with no header image
A new page for the Prison Officer leg was needed.
This page wouldn't have a header image,
or a link in the leg nav.

As leg pages expect an image to be present in order for the layout to work,
a spacer element has been added to ensure that there is still enough
room in the header for the page title.

This new page type should still go through the normal design/dev process at some point.