Re-enable find-a-job page job feed message
This is primarily used to create a banner with a link
to re-direct users who land on the Youth Custody find a job page
to the Prison Officer find a job page,
as there tend to be fewer jobs on the find-a-job page.

The values for this are currently hardcoded.
They used to be user editable.
This is being re-enabled.