Populate jobListMessage by WYSIWIG field
Commented out functionality to add search terms to job list message link.
The job list message is now populated by a WYSIWYG field
and so is not guaranteed to be a link.

A more sophisticated solution will be needed to ensure that any links to other
find a job pages will have the correct search terms applied.