Turn off font ligatures
Users are currently reporting an issue where 'ffi' isn't displaying. This is quite a prominent issue, as the word officer is used a lot across the site. I suspect this is caused by the font not having a ligature available for 'ffi', and so in its absense, it's not rendering anything.

There is a chance this is a browser bug that was introduced in Firefox 75, and this will be replaced by a new version soon. In the meantime though, this should fix the issue for users.