Pull in cookie banner and modify theme to accept it
This pulls in the cookie banner. It also makes some changes to the base theme to allow it to work properly. These will want to come out at a later date, when we've made the cookie banner independent of base theme settings, and encapsulate the CSS in it. I'll put a note about this on a Trello card. But for now, this gets it onto the PPJ site for our WCAG audit next week.